+++ src/processmanager.h
@@ -10,6 +10,8 @@ class ProcessManager : public QObject {
   Q_PROPERTY(QString output READ output NOTIFY outputChanged)
   Q_PROPERTY(bool isRunning READ isRunning NOTIFY isRunningChanged)
   Q_PROPERTY(bool isPaused READ isPaused NOTIFY isPausedChanged)
+  Q_PROPERTY(int lastExitCode READ lastExitCode NOTIFY lastExitCodeChanged)
+  Q_PROPERTY(bool hasFinished READ hasFinished NOTIFY hasFinishedChanged)
 
 public:
   explicit ProcessManager(QObject *parent = nullptr);
@@ -18,6 +20,8 @@ public:
   QString output() const { return m_output; }
   bool isRunning() const { return m_isRunning; }
   bool isPaused() const { return m_isPaused; }
+  int lastExitCode() const { return m_lastExitCode; }
+  bool hasFinished() const { return m_hasFinished; }
 
   Q_INVOKABLE void runCommand(const QString &program,
                               const QStringList &arguments = QStringList());
@@ -37,6 +41,8 @@ signals:
   void outputChanged();
   void isRunningChanged();
   void isPausedChanged();
+  void lastExitCodeChanged();
+  void hasFinishedChanged();
   void commandFinished(int exitCode, const QString &output);
 
 private slots:
@@ -53,6 +59,8 @@ private:
   QString m_output;
   bool m_isRunning;
   bool m_isPaused;
+  int m_lastExitCode;
+  bool m_hasFinished;
 };
 
 #endif // PROCESSMANAGER_H